DE-2377 - Backport to Java 8
In our current strimzi based setup, we are still using java 8, hence we
need this connector to be binary compatible with that VM.

Relaxed compilation parameters from warnings to errors, as many of them
we can ignore them by the moment.

Also disabled javadoc for the moment.

As a summary of changes (from big to little):

* loads of var replacements to current types
* loads of List.of replaced with Arrays.asList
* loads of Map.of replaced with "new HashMap<>() {{ }}" pattern
* backported String.isBlank() with a regex